fix: splitfile
